chinook: now handles custom sleep syntax in cmd_script
Custom sleep times can now be appended to commands in cmd_history (for
auto_next mode only)

If '!2000' is found at the end of a command line, the script will pause
for 2 seconds after that command is entered. This has allowed adding
custom pause times for a few select commands that take a little longer
to run. With this feature plus the breaking from auto_next, auto_next
can now be used for recording instead of just testing...