chinook: macro sections now scroll to positions + misc
Turned off the "jump to" macro commands and replaced with literal
Up/Down sequences so the user can see the continuous cursor position.
The downside of this is that any changes mean all the subsequent offsets
have to be manually recalculated, but I don't expect any more changes to
the macros so this shouldn't be an issue.

Also tweaked some of the cmd_Script comment text